test(unread): extract shared test infrastructure; consolidate stale-fence pair
Extract DOM shim, localStorage helpers, event factories, and mount
harnesses into observedUnreadTestHarness.mjs, a non-test helper module
(not matched by src/**/*.test.mjs). All three hook test files import
from it instead of copy-pasting the infrastructure.

Consolidate the stale clearAll and stale removeChannel scope-fence
tests (useObservedUnreadPersistence) into one test that asserts both
operations in sequence — failure messages name the specific operation.
All assertions from both original tests are preserved.

Zero production changes. Zero assertions deleted.
Test-file line count: 1,866 → 1,527 (-339 lines in .test.mjs files).
