+++ desktop/src/features/channels/useUnreadChannels.test.mjs
@@ -12,275 +12,23 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-// ── DOM shim ─────────────────────────────────────────────────────────────────
-// Same shim used by useObservedUnreadPersistence.test.mjs: minimal DOM with
-// full event-listener support so createRoot + pagehide flush both work.
+import {
+  installDOMShim,
+  installFreshStorage,
+  seedStorage,
+  mountUnreadChannels,
+} from "./observedUnreadTestHarness.mjs";
 
-function installDOMShim() {
-  class ET {
-    constructor() {
-      this._ls = {};
-    }
-    addEventListener(t, fn) {
-      this._ls[t] ??= [];
-      this._ls[t].push(fn);
-    }
-    removeEventListener(t, fn) {
-      this._ls[t] = (this._ls[t] ?? []).filter((f) => f !== fn);
-    }
-    dispatchEvent(e) {
-      for (const fn of this._ls[e.type] ?? []) fn(e);
-      return true;
-    }
-  }
-  class Node extends ET {
-    constructor(tag) {
-      super();
-      this.tagName = tag;
-      this.children = [];
-      this.childNodes = [];
-      this.style = {};
-      this.nodeType = 1;
-      this.parentNode = null;
-    }
-    get ownerDocument() {
-      return globalThis.document;
-    }
-    get firstChild() {
-      return this.children[0] ?? null;
-    }
-    get lastChild() {
-      return this.children[this.children.length - 1] ?? null;
-    }
-    get nextSibling() {
-      return null;
-    }
-    get nodeValue() {
-      return null;
-    }
-    appendChild(c) {
-      this.children.push(c);
-      this.childNodes.push(c);
-      c.parentNode = this;
-      return c;
-    }
-    removeChild(c) {
-      this.children = this.children.filter((x) => x !== c);
-      this.childNodes = this.childNodes.filter((x) => x !== c);
-      return c;
-    }
-    insertBefore(n, r) {
-      if (!r) return this.appendChild(n);
-      const i = this.children.indexOf(r);
-      if (i < 0) return this.appendChild(n);
-      this.children.splice(i, 0, n);
-      this.childNodes.splice(i, 0, n);
-      n.parentNode = this;
-      return n;
-    }
-    contains(n) {
-      return this === n || this.children.some((c) => c?.contains?.(n));
-    }
-  }
-  class Doc extends ET {
-    constructor() {
-      super();
-      this.nodeType = 9;
-    }
-    createElement(t) {
-      return new Node(t);
-    }
-    createTextNode(v) {
-      const n = new Node("#text");
-      n.nodeType = 3;
-      n.nodeValue = v;
-      return n;
-    }
-    createComment(v) {
-      const n = new Node("#comment");
-      n.nodeType = 8;
-      n.nodeValue = v;
-      return n;
-    }
-    get activeElement() {
-      return null;
-    }
-    contains(n) {
-      return n != null;
-    }
-  }
-
-  const windowET = new ET();
-  globalThis.document = new Doc();
-  globalThis.HTMLIFrameElement = Node;
-  globalThis.HTMLElement = Node;
-  globalThis.IS_REACT_ACT_ENVIRONMENT = true;
-  process.env.IS_REACT_ACT_ENVIRONMENT = "true";
-  if (typeof globalThis.window === "undefined") {
-    Object.defineProperty(globalThis, "window", {
-      value: globalThis,
-      configurable: true,
-    });
-  }
-  globalThis.addEventListener = windowET.addEventListener.bind(windowET);
-  globalThis.removeEventListener = windowET.removeEventListener.bind(windowET);
-  globalThis.dispatchEvent = windowET.dispatchEvent.bind(windowET);
-  globalThis.MutationObserver = class {
-    observe() {}
-    disconnect() {}
-    takeRecords() {
-      return [];
-    }
-  };
-  globalThis.requestAnimationFrame = (fn) => setTimeout(fn, 0);
-}
-
+// DOM shim must run before any React import (harness imports React at parse time).
 installDOMShim();
-
-// ── localStorage shim ─────────────────────────────────────────────────────────
-
-function makeLocalStorage() {
-  const store = new Map();
-  return {
-    get length() {
-      return store.size;
-    },
-    key: (i) => [...store.keys()][i] ?? null,
-    getItem: (key) => store.get(key) ?? null,
-    setItem: (key, value) => store.set(key, value),
-    removeItem: (key) => store.delete(key),
-    clear: () => store.clear(),
-    _store: store,
-  };
-}
-
-function installFreshStorage() {
-  const ls = makeLocalStorage();
-  Object.defineProperty(globalThis, "localStorage", {
-    get: () => ls,
-    configurable: true,
-  });
-  return ls;
-}
-
-// Install once at module level.
 installFreshStorage();
 
-// ── Imports (after shims) ─────────────────────────────────────────────────────
-
-import React from "react";
-import { createRoot } from "react-dom/client";
+import { readObservedUnreadFromStorage } from "./observedUnreadStorage.ts";
 import { act } from "react";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-
-import { useUnreadChannels } from "./useUnreadChannels.ts";
-import {
-  readObservedUnreadFromStorage,
-  writeObservedUnreadToStorage,
-} from "./observedUnreadStorage.ts";
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
 
 const RELAY = "wss://relay.example.com";
-// Use a recent timestamp so age-pruning doesn't drop events before we assert.
-// 7-day horizon: events must be newer than (now - 604800 seconds).
-const NOW_S = Math.floor(Date.now() / 1_000);
-
-function makeObservedEvent(id, _channelId, createdAt = NOW_S) {
-  return {
-    id,
-    createdAt,
-    rootId: `root-${id}`,
-    highPriority: false,
-    countsTowardBadge: true,
-    countsTowardAppBadge: false,
-  };
-}
-
-/**
- * Pre-seed storage for a given pubkey+relay with one channel's worth of events.
- * Returns the ISO timestamp matching the seeded event's createdAt, suitable for
- * passing as the `readAt` argument to markChannelRead to trigger clearObserved.
- */
-function seedStorage(pubkey, relay, channelId, eventId = "evt-1") {
-  const inner = new Map();
-  inner.set(eventId, makeObservedEvent(eventId, channelId, NOW_S));
-  const events = new Map();
-  events.set(channelId, inner);
-  writeObservedUnreadToStorage(pubkey, relay, events);
-  // Return a timestamp that covers the seeded event (equals createdAt).
-  return new Date(NOW_S * 1_000).toISOString();
-}
-
-/**
- * Mount useUnreadChannels inside a QueryClientProvider harness.
- * Returns controls to inspect and re-render the hook.
- */
-async function mountUnreadChannels({ pubkey, relay = RELAY }) {
-  const qc = new QueryClient({
-    defaultOptions: { queries: { retry: false } },
-  });
-
-  let capturedMarkChannelRead = null;
-  let capturedMarkAllChannelsRead = null;
-
-  function Inner({ pubkey: pk }) {
-    const result = useUnreadChannels([], null, {
-      pubkey: pk,
-      // relayClient prop is undefined — useReadState returns no-ops (no relay
-      // calls). The module-level relayClient used by useLiveChannelUpdates is
-      // the real singleton but operates on an empty channel list, so no
-      // subscriptions are attempted.
-      relayClient: undefined,
-      relayUrl: relay,
-    });
-    capturedMarkChannelRead = result.markChannelRead;
-    capturedMarkAllChannelsRead = result.markAllChannelsRead;
-    return null;
-  }
-
-  function Harness({ pubkey: pk }) {
-    return React.createElement(
-      QueryClientProvider,
-      { client: qc },
-      React.createElement(Inner, { pubkey: pk }),
-    );
-  }
-
-  const container = document.createElement("div");
-  const root = createRoot(container);
-
-  const render = async (pk) => {
-    await act(async () => {
-      root.render(React.createElement(Harness, { pubkey: pk }));
-    });
-  };
-
-  await render(pubkey);
-
-  return {
-    get markChannelRead() {
-      return capturedMarkChannelRead;
-    },
-    get markAllChannelsRead() {
-      return capturedMarkAllChannelsRead;
-    },
-    render,
-    unmount: async () => {
-      await act(async () => root.unmount());
-    },
-    /** Fire a pagehide event to synchronously flush any pending debounce. */
-    flushStorage: () => {
-      dispatchEvent(
-        new (class extends Event {
-          constructor() {
-            super("pagehide");
-          }
-        })(),
-      );
-    },
-  };
-}
 
 // ── Tests ─────────────────────────────────────────────────────────────────────
